Suppress inclusion of Cuda stubs at runtime
Nix can be immensely frustrating. If you run into "CUDA driver is a stub library", this changeset works around it by killing all stub files during compilation and insisting on the real ones. Sadly this is inimical to caching, but it's good enough to be a stopgap with Nix's troublesome state of affairs.

If you are not running on NixOS, you must link /usr/lib/x86_64-linux-gnu/{libcu*,libglut*,libglib*,libglapi*} into a new directory at /run/opengl-driver/lib for this to auto-detect it without LD_LIBRARY_PATH.